+++ modules/tinymce/src/themes/silver/main/ts/ui/core/complex/FontFamilyBespoke.ts
@@ -4,6 +4,7 @@ import { Arr, Fun, Optional, Optionals } from '@ephox/katamari';
 import Editor from 'tinymce/core/api/Editor';
 
 import * as Events from '../../../api/Events';
+import * as Options from '../../../api/Options';
 import { UiFactoryBackstage } from '../../../backstage/Backstage';
 import { updateMenuText } from '../../dropdown/CommonDropdown';
 import { onSetupEditableToggle } from '../ControlUtils';
@@ -16,6 +17,10 @@ const systemFont = 'System Font';
 
 // A list of fonts that must be in a font family for the font to be recognised as the system stack
 // Note: Don't include 'BlinkMacSystemFont', as Chrome on Mac converts it to different names
+// The system font stack will be similar to the following. (Note: each has minor variants)
+// Oxide: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+// Bootstrap: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", Arial, "Noto Sans", sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol", "Noto Color Emoji";
+// Wordpress: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen-Sans, Ubuntu, Cantarell, "Helvetica Neue", sans-serif;
 const systemStackFonts = [ '-apple-system', 'Segoe UI', 'Roboto', 'Helvetica Neue', 'sans-serif' ];
 
 // Split the fonts into an array and strip away any start/end quotes
@@ -24,33 +29,31 @@ const splitFonts = (fontFamily: string): string[] => {
   return Arr.map(fonts, (font) => font.replace(/^['"]+|['"]+$/g, ''));
 };
 
-const isSystemFontStack = (fontFamily: string): boolean => {
-  // The system font stack will be similar to the following. (Note: each has minor variants)
-  // Oxide: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-  // Bootstrap: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", Arial, "Noto Sans", sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol", "Noto Color Emoji";
-  // Wordpress: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen-Sans, Ubuntu, Cantarell, "Helvetica Neue", sans-serif;
-  const matchesSystemStack = (): boolean => {
-    const fonts = splitFonts(fontFamily.toLowerCase());
-    return Arr.forall(systemStackFonts, (font) => fonts.indexOf(font.toLowerCase()) > -1);
-  };
+const matchesStack = (fonts: string[], stack: string[]): boolean => Arr.forall(stack, (font) => fonts.indexOf(font.toLowerCase()) > -1);
 
-  return fontFamily.indexOf('-apple-system') === 0 && matchesSystemStack();
+const isSystemFontStack = (fontFamily: string, userStack: string[]): boolean => {
+  if (fontFamily.indexOf('-apple-system') === 0 || userStack.length > 0) {
+    const fonts = splitFonts(fontFamily.toLowerCase());
+    return matchesStack(fonts, systemStackFonts) || matchesStack(fonts, userStack);
+  } else {
+    return false;
+  }
 };
 
 const getSpec = (editor: Editor): SelectSpec => {
-
   const getMatchingValue = () => {
     const getFirstFont = (fontFamily: string | undefined) => fontFamily ? splitFonts(fontFamily)[0] : '';
 
     const fontFamily = editor.queryCommandValue('FontName');
     const items = dataset.data;
     const font = fontFamily ? fontFamily.toLowerCase() : '';
+    const userStack = Options.getDefaultFontStack(editor);
 
     const matchOpt = Arr.find(items, (item) => {
       const format = item.format;
       return (format.toLowerCase() === font) || (getFirstFont(format).toLowerCase() === getFirstFont(font).toLowerCase());
     }).orThunk(() => {
-      return Optionals.someIf(isSystemFontStack(font), {
+      return Optionals.someIf(isSystemFontStack(font, userStack), {
         title: systemFont,
         format: font
       });